refactor top-level into coreplex and platform
