+++ src/main/scala/TestConfigs.scala
@@ -69,7 +69,10 @@ class WithComparator extends Config(
     case BuildGroundTest =>
       (p: Parameters) => Module(new ComparatorCore()(p))
     case ComparatorKey => ComparatorParameters(
-      targets    = Seq("mem", "io:int:testram").map(name => site(GlobalAddrMap)(name).start.longValue),
+      targets    = Seq(
+        site(GlobalAddrMap)("mem"),
+        site(ExtAddrMap)("testram"))
+          .map(entry => entry.start.longValue),
       width      = 8,
       operations = 1000,
       atomics    = site(UseAtomics),
